Fixed S3M channel panning
Reworking a code you wrote ten years ago when you were learning C is
always a pain. This time I really didn't understand what I wanted to
do with a very bizarre code in the S3M loader, so I rewrote it based
on the specs. Channel panning seems to be better now, but it will need
some testing.