Implement Protracker-style sample loops
Protracker plays sample loops with loop end before the sample end this
way: first play the entire sample, then play the loop. MeNoWantMiseria.mod
needs this feature to be implemented to play correctly.

Note: this works only on the software mixer. The OSS sequencer API does
not implement this kind of loop, nor do the GUS and AWE cards.