API Addition: xmp_set_row
This patch adds code, documentation, and unit testing for a new function, xmp_set_row, as discussed in issue #100.  I must confess, however, that I do not fully understand the purpose of the final if block in set_position, which is the reason I had to make a special case for a negative pos member of player_data.

Please review and share your thoughts.  Thank you for maintaining libxmp.