Fix memory access issues playing smix samples
Our mixer routine is a large chunk of legacy code written in the 1990s
trying to save as much CPU as possible. It kind of succeeded, but the
code itself is messy and nearly unmaintanable. When a bug appears, the
fix usually propagates ugliness across the code, and this one is no
exception.

The interpolators we use here read previous values directly from the
sample data instead of keeping a small internal queue, so we add a
small empty space before the sample data. When external RIFF samples
were used in smix, this extra space was not added, leading to memory
access problems being pointed by valgrind.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>