Reset buffer state on player start
When playing in "fill the buffer" mode, there's an internal that contains
a pointer to our mixing buffer. If the mixer is disabled and the replay
is resumed with a new mixer session, the state may become inconsistent.
We'll now force state reset at player start to prevent this problem.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>