Fix MEDs with pattern lengths >256, MED instrument corruption.
Adds a new function to allow MEDs with block lengths >256 to use
a relaxed row limit and adds MED-specific sanity checks for block
lengths (3200 for Amiga MEDs, 9999 for MED Soundstudio 2).

Also fixes an issue with MED instrument name loading that would cause
the instrument volume and number of samples to get corrupted.