loaders: fix loading xm instruments with more than 16 samples
The FT2 specs state that the limit of samples per instrument is 16,
however there are XM modules that exceed this limit (for example,
grass near the house.xm has an instrument with 23 samples). Increase
the limit to 32 to cover these cases, but if a larger number is
found we should handle this in a more dynamic way.

Fixes #168

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>