loaders: increase number of instruments in STM header to 32
STM v1 may define 32 instruments instead of the 31 instrument limit in
the STM v2 header, so add an extra instrument slot and check if we're
not exceeding this number before loading data.

It's also interesting to note that this problem was not detected in
the loader tests running on Linux (but they crashed on osx).

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>