player: fix volume slide lag
AliceLR reports that effect processing for volume and pitch slides are
lagging behind one frame, and that's caused by the relevant control
variables being updated after the current volume and pitch values are
already sent to the sample mixer. This results in changes that should
happen in the first frame of each row actually becoming effective in
the second frame.

To fix that, we're now updating the slide variables before using them.
This caused a number of regression tests to fail, and those were
updated to use the correct reference values.

Fixes #242, #260

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>