player: read row events only once per row
IT fine volume slides should be applied only once per row even if the
frame counter resets to 0 due to a pattern delay (IT effect SEx).
We're using bit 1 of the rowdelay_set flag (which is a bit confusing,
perhaps we should refactor this or change the names) to mark the
real first frame of this row, so also verify it when reading new
row events.

Fixes #264

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>