Remove 'quirk87' handling from the S3M loader.
This compatibility check was added to fix a broken version of
"Return of Litmus" that existed circa 2007. This broken version
no longer seems to be available, and all copies I can find are
not corrupted. Furthermore, support for this hack was broken
anyway due to several sanity checks on the affected variables
having been added. The previously reported behavior of libxmp
hanging at 100% CPU usage when loading this broken module should
no longer occur.