Add DJGPP configure.ac fork patch, move decrunch to src/depackers/.
This moves the functions `decrunch`->`libxmp_decrunch` and
`libxmp_exclude_match` to a new compilation unit in src/depackers/.
This isolates all `execute_command` kludges as well as xfnmatch.c to
the depackers subsystem instead of in core code.

This also applies sezero's configure.ac patch to prevent DJGPP's
dummy fork/exec/etc. functions from incorrectly causing the fork
version of `execute_command` to be selected for this platform.

Co-authored-by: sezero <sezero@users.noreply.github.com>