Fix Impulse Tracker envelope and fadeout order.
IT instruments initiate fadeout when the end of their envelope is
reached. This should apply immediately, but was delayed by multiple
ticks by the order of envelope tick -> check for envelope end ->
apply fadeout being reversed.

Instead of being delayed by 3 ticks relative to Impulse Tracker,
the fadeout is now only delayed by 1 tick. This remaining delay
is due to the envelope end check allowing the last envelope tick
to play before starting the fadeout, which may be preferable despite
not being 100% accurate to IT (subjective).