IT new inst. fade envelope reset should only affect volume envelope.
When a new instrument is encountered and the IT volume envelope has
started fadeout, the volume envelope should be reset, but libxmp
would reset all three envelopes in this situation. The pan and pitch
envelopes should not be affected by this check--when they don't have
carry set, they should reset independently of this situation, and
when they do have carry set, they should not be modified.

Verified with Impulse Tracker 2.14p5 and OpenMPT 1.29.10.