Don't increment voice position by step value at loop/tick end.
Since commit b7d6a6ea, libxmp has added an extra step to voice
positions if they are close to or at the end of their loop on a tick
boundary. This results in samples with loops being very slightly
too fast. This patch corrects this behavior so samples with loops
play at the correct speed.

The issue can be demonstrated with two looping samples, where one
sample is two inverted copies of the other sample. When played back
at the same pitch, they should cancel entirely and be silent, but
this extra step causes them to gradually diverge instead.