Fix for IT filter cutoff 127 on new note behavior.
The IT filter should be disabled if the channel filter cutoff is
currently 127 and the channel filter resonance is currently 0. This
combination of settings should only be applied when a new note is
encountered that does not use portamento; otherwise, the previous
settings should be used. libxmp was instead always resetting the
cutoff to 127 on a new note.