Fix NNA and tone portamento interaction with sample changes.
When tone portamento is used on an instrument with an NNA and
the old note and the new note have different samples, the
tone portamento should cancel the NNA and duplicate check for
that note but should not actually be played.

libxmp was canceling the NNA correctly before, but it was clearing
the entire mixer voice instead of just the current NNA state. libxmp
was NOT canceling the duplicate check before in this situation, which
allowed the portamento note to incorrectly cancel past notes.

test/test.c change is due to an extra anticlick decay being generated
due to not clearing its state anymore.