Implement S3M and IT mix volume.
S3M and IT both contain a mix volume parameter, which ranges from
16 to 128 (S3M), 0 to 128 (IT), or 0 to 255 (OpenMPT), and typically
has a default value of 48. The way this value is applied varies but
a common approach is to use (mix volume)/256 to reduce the level of
a -0dB sample. Since libxmp effectively quarters the level of samples
by default through other means, this implementation just treats 48 as
the "base" and allows higher values to apply gain for now.

Other changes made by this patch:
* Removed a sanity check for IT mix volume--OpenMPT can use the full
  range of this field.
* Revised S3M channel default panning calculation, which seems to
  have not been especially grounded in reality.
* Added old S3M mix volume conversion and adjustment for mono modules.