Rename "Protracker Studio" loader to "Epic MegaGames MASI 16".
There is no such thing as "Protracker Studio"/"ProTracker Studio".
While this was technically the name of the format, these were created
using a converter that was part of Epic's MASI sound system.

The 16-bit MASI variant, aka "PS16" or "PSM16", is just an older
version of the same format that is handled by libxmp's MASI loader.
Having them named completely differently falsely implies that they
are unrelated; the 16-bit variant is just the older version of the
format, used in the Epic Pinball shareware disks and Silverball.