Add compatibility for Modplug Tracker preamp scaling.
Modplug Tracker and early versions of OpenMPT have a pre-amp routine
that, among other things, attenuates the mix volume based on the
current module's channel count. This channel scaling needs to be
implemented in libxmp to fix the output levels of some especially
loud MPT modules.

Modplug Tracker will scale the pre-amp level "logarithmically" from
100% (<=5 channels) to 50% (30+ channels) at the default setting 128.
(This changes for higher settings, but not lower settings.)

Currently this is implemented in the loaders instead of the mixer,
and only the default setting of 128 is supported. ITs should work
fine now, but XMs need better MPT detection to be added.