Fix an extra scan edge case related to scan edge case fix #688. (#692)
If two sequences jump to the same invalid/skip pattern immediately
prior to the end of the module, the previous scan fix patch would
attempt to read the order list out-of-bounds. I've fixed this by
just making the invalid/skip pattern handling part of the same
branch as the S3M_END handling in the fix, which I should have done
anyway.