Faster IT loading by buffering pattern, sample, and comment reads. (#694)
The IT format stores 16-bit lengths for patterns and compressed sample
blocks, which allows both to be loaded into a buffer to be read from
memory. This patch adds a 64k temporary buffer to the IT loader to
reduce load times for this format.

The average improvement in my loose measurements is about 45.5%
(measured by loading every file in the Modland IT directory once with
xmpchk: Windows 10 x86_64, all files buffered by the kernel, all files
read into memory and loaded with `xmp_load_module_from_memory`).
This is going to vary based on the structure of any given file.

|       | master (16fc5c89) | Patch      |
|-------|-------------------|------------|
| Run 1 | 3m34.269s         | 1m56.595s  |
| Run 2 | 3m34.537s         | 1m56.572s  |
| Run 3 | 3m34.021s         | 1m57.067s  |
| Avg.  | 3m34.276s         | 1m56.745s  |

Further testing in Fedora and some additional testing by ds-sloth on
the 3DS verifies these performance improvements.