Replace rand() with a built-in reentrant alternative. (#728)
This patch replaces all rand() usage with a built-in implementation
of xorshift32. This is not an especially cryptographically good
PRNG but it is fast (4 bitshifts, 3 XORs, 1 multiply for output)
and produces reasonably "random" output.

Unlike rand(), this implementation is reentrant i.e. it easily
supports multiple different concurrent states and doesn't cause
any race conditions with caller code outside of libxmp.

This implementation will also behave exactly the same on all
platforms given a fixed seed, which allows the fuzzer/xmpchk and
the storlek_20.it test to use fixed seeds. (Lack of doing this has
caused failures of this test on NetBSD and in neumatho's C# port.)

Fixes #685 and #723.