Increase libxmp maximum sample rate to 768kHz.
Increases XMP_MAX_SRATE to 768000, which in turn increases
XMP_MAX_FRAMESIZE to 384000. To compensate for this, the tick buffers
are allocated using the player-provided rate (which requires a call
to xmp_start_player to change) instead of with XMP_MAX_FRAMESIZE.
I've added a warning to libxmp.rst that the frame info buffer
may also be much smaller than XMP_MAX_FRAMESIZE.

In case of existing software misusing the old XMP_MAX_FRAMESIZE,
the minimum value for s->total_size is the old XMP_MAX_FRAMESIZE
for now.