Add depacker test_hio function for more complex packed formats. (#750)
This allows depackers to test the input stream directly rather than
the 1k buffer that libxmp currently provides them, which will be
useful for formats that don't provide a magic string at the start.

Example 1: Pack-Ice--a packer format used on the Atari ST and Falcon
(and occasionally the Amiga?)--in the 1.x variant of its format
writes a footer with a 4 byte signature at the end of the file.
It is impossible to detect this variant otherwise. I am currently
finishing a depacker for the Pack-Ice format because Digital Tracker
natively loads the 2.3/2.4 variant of it (allegedly; unpacking seems
broken prior to Digital Home Studio) and the author regularly used
Pack-Ice 2.3/2.4 on their modules.

Example 2: Zip does not reliably contain a magic string at the start
(our Zip depacker currently assumes there is). A single volume Zip
file can be more reliably detected by seeking to END - 22 and then
searching backwards up to 65535 bytes for the Zip End of Central
Directory record. This is because, due to the nature of the Zip
format, it can be prefixed with an arbitrary-sized amount of data
("embedded stub"). Zip self-extractors may rely on this, but most
Zip files don't, so I don't think there's any need to rush to
implement this.