Merge song file instrument path detection routines.
I found a few glaring issues that needed addressing while fixing
a MED2 loader snprintf truncation warning related to sample loading.
The MOD, STM, MED2, and (bitrot) SSMT loaders all contain separate
routines to generate the path of an instrument/sample file when
loading song files. These are implemented roughly as follows:

* MOD: combine `m->dirname` with the cleaned sample name, open.
* STM: combine `m->dirname` with the cleaned sample name, open.
* SSMT: combine `m->dirname` with the cleaned sample name, open.
* MED2:
  1) Get the instrument path set by `xmp_set_instrument_path` or if
     it is not set, get the environment variable
     `XMP_INSTRUMENT_PATH`, or if it is not set, `.`. Combine this
     with the cleaned sample name and attempt a case-insensitive
     match, then open.
  2) Combine `m->dirname` with the cleaned sample name, attempt a
     case-insensitive match, then open. This was added by #416 for
     consistency with the other loaders.

Clearly, one of these is not like the others! There are a few major
issues here.

1) The MOD, STM, and SSMT loaders totally ignore
   `xmp_set_instrument_path` and/or `XMP_INSTRUMENT_PATH`, which
   sucks and means this feature is only useful for MED2, a format
   where literally no test files existed and I had to make my own
   for libxmp's regression tests. Note that this feature allows
   memory modules to load instruments, and `m->dirname` only works
   for file modules.
2) The MOD, STM, and SSMT loaders fail to check for instrument files
   case-insensitively. MOD is from the same platform as MED2 and STM
   is explicitly from a case-insensitive operating system, so this is
   a strange omission.
3) MED2, on the other hand, attempts to check the current directory
   if neither `xmp_set_instrument_path` or `XMP_INSTRUMENT_PATH` is
   set, which is strange and seems like a worse version of loading
   from `m->dirname`. Worse, if this was applied to MOD/STM, it could
   cause new misbehavior that `xmp_set_instrument_path` and
   `XMP_INSTRUMENT_PATH` are less likely to, so I removed it.
   Intentional "regressions" to MED2 are better than to MOD/STM.
4) No failure checks on any of the `snprintf` occurrences, so really
   these `fopen` calls could be opening anything if they truncate.

The new, combined handling for all loaders is now:

1) Get `xmp_set_instrument_path` or `XMP_INSTRUMENT_PATH` if it is
   set, combine it with the cleaned sample name, and attempt a
   case-insensitive search. If all the copies were successful and
   the file was found, open.
2) Get `m->dirname` if it is set, combine it with the cleaned sample
   name, and attempt a case-insensitive search. If all the copies
   were successful and the file was found, open.

TANGENT TIME: `strncpy` is the worst C standard library function.
`strncpy` will pad its entire buffer with `\0` even when this isn't
wanted or requested, but when the input string is longer than the
output buffer, it will not null terminate the output buffer. It is
not safe. After merging the handlers to address the above issues,
GCC's stringop analyzer found several misuses of this function that
have been hiding in plain sight for years and opening us up to more
out-of-bounds reads. I've replaced them all with `snprintf` as this
also allows the detection of truncation with its return value, which
in turn can be carried through as a fail state for sample loading.