Clean up jank in mixer.c sample loop handling.
mixer.c was handling sample loops in two different places:

1) If the end of the tick exactly coincided with the channel
   reaching the end of its loop, it would be looped immediately
   before the mixer continued to the next channel.
2) If the end of the sample was reached prior to the end of the
   tick, it would be looped after checking for a one-shot sample.

This originally wasn't too horrible, but the addition of reverse
samples and resampling wraparound management bloated it. Fixing
Protracker sample swaps (#717) is going to bloat it further.
These two cases were mostly equivalent, so I combined the former
into the latter.

Oktalyzer split channel support is affected by this, but most
likely the change is a net positive. When a sample plays in an
Oktalyzer split channel, it is not supposed to loop, but case 1
above was allowing sample loops to leak through if they
coincided with the end of a tick. This no longer occurs.