Implement support for Protracker instrument swapping.
Protracker 1/2 events that contain an instrument number but no
note (or a note and toneporta) behave different from most other
MOD players; the channel volume, finetune (though not necessarily
period), and sample pointer are set immediately, but the sample
change does not take effect until the currently playing sample
reaches the end of its loop. I've implemented support for this--at
least as far as OpenMPT's tests cover--under the QUIRK_PROTRACK
quirk, which required significant changes to read_event_mod and
some new mixer functionality. libxmp now passes 7 of the 9 related
OpenMPT tests that are not part of the regression tests (previously
only 1); I've added these to the regression tests.