Fix DBM division-by-zero from bad bounding during pattern load.
The DBM pattern loader was checking for end-of-pattern AFTER loading
bytes from the pattern, rather than before, leading to cases where
it would load an effect and then fail to convert it. This was
allowing FX_ICE_SPEED to be loaded for DBMs with a parameter that
causes the player to crash.

The DBM pattern loader now checks for the end of the pattern BEFORE
loading bytes from the pattern, and reaching the end of the pattern
in the middle of an event will no longer prevent effects conversion.

(I also combined two DBM fuzzer truncation tests to reduce the
linker command length slightly.)