Scan should only count one instance of pattern delay per row.
Only one pattern delay is ever executed per row, but the module
scan was counting every instance of pattern delay, leading to
wildly inaccurate times in some cases (bugfixed dammed illusion.mod
should be 5m55s, but it was being counted as 6m05s). This
affected the test data for OpenMPT PatternDelays.xm, which is still
wrong due to the other effect it tests being missing.

See also the IT equivalent of this issue, since it has its own
pattern delay implementation for some reason:
https://github.com/libxmp/libxmp/issues/808