Fix rare mixer crash when reversing a sample past its end. (#819)
In some cases with a low player sample rate and a sample
playing at a very high rate, the mixer can advance the current
sample position well past the end of the sample. When this was
followed by a reverse sample effect e.g. IT S9F, the mixer
would try to read the sample out of bounds. I'm not convinced
this is 100% fixed but I wasn't able to get it to reproduce
with my own modules.