Add support for loading MED4 song files, fix MED4 hybrids.
MED4 external instruments, unlike MED2 and MED3, can be a sample,
a synthsound, or a hybrid. Due to this and other differences between
MED4 and the other formats, MED4 needs custom implementations of
almost everything to fully allow song files to work.

This isn't a small overhaul, so to summarize what I did:

* Split the MED4 sample instrument loader out to its own function
  and gave it the ability to reallocate the samples list.
* Merged the MED4 synth and hybrid instrument loaders into a single
  function and gave it the ability to reallocate the samples list.
  This fixes hybrid instruments with multiple waveforms.
* Added a song file samples loader for MED4 which doesn't rely on
  the samples mask (which isn't saved in a song file).
* Hex edited several fuzzer regression test to set bit 3 of the
  song flags so they would work as intended.

Like MED3, a song file can be detected by checking for bit 3 of
the song flags (0=song, 1=module).