Fix array subscript underflow in Pha Packer loader (CVE-2025-47256).
The array iterator k can be fed a garbage initial value that allows
it to overflow to INT_MIN - 1 in some circumstances. This, paired
with truncated modulo being used instead of AND, leads to array
underflows of onote and ocpt. The initial value of k doesn't actually
matter, so I've changed it to 0 to match the last upstream I have
available.