Add Software Visions DMF support to the MOD loader. (#862)
Adds support for the variant of MOD used with the DOS game
Apocalypse Abyss and for the Software Visions 1995 Catalog. These
are just ModEdit M.K. MODs with the extension .DMF, the title and
sample names stripped, and the first 2108  bytes word-flipped to
little endian (either as a weak obfuscatory technique or to make
loading the samples slightly faster).

I was originally requested and was going to adapt the ProWizard
SV.C converter for this, but after reviewing the MOD loader I
realized the required changes aren't that bad. I haven't measured
any perceivable change in testing/loading speed.

(There is a converter in the wild that incorrectly swaps 2104
bytes instead of 2108. I used FROZEN.DMF for the test module as a
check on this in case there's a well-intentioned attempt to "correct"
this in the future, since it's one of the two Apocalypse Abyss
modules to use the last event of the first pattern.)