Remove kernel list.h usage in iff.c, other list.h references in libxmp.
This header generates several annoying and otherwise unfixable warnings
for KallistiOS builds for me downstream:

-----

In file included from contrib/libxmp/src/loaders/iff.c:24:
contrib/libxmp/src/loaders/../list.h:22: warning: "LIST_HEAD" redefined
   22 | #define LIST_HEAD(name) \
      |
In file included from /opt/toolchains/dc/kos/include/kos/tls.h:25,
                 from /opt/toolchains/dc/kos/include/kos/thread.h:43,
                 from /opt/toolchains/dc/kos/include/sys/sched.h:28,
                 from /opt/toolchains/dc/sh-elf/sh-elf/include/sys/_pthreadtypes.h:23,
                 from /opt/toolchains/dc/sh-elf/sh-elf/include/sys/types.h:221,
                 from /opt/toolchains/dc/sh-elf/sh-elf/include/sys/unistd.h:12,
                 from /opt/toolchains/dc/sh-elf/sh-elf/include/unistd.h:4,
                 from /opt/toolchains/dc/kos/include/kos/dbglog.h:25,
                 from /opt/toolchains/dc/kos/include/sys/stdio.h:30,
                 from /opt/toolchains/dc/sh-elf/sh-elf/include/stdio.h:85,
                 from contrib/libxmp/src/loaders/../common.h:27,
                 from contrib/libxmp/src/loaders/iff.c:23:
/opt/toolchains/dc/sh-elf/sh-elf/include/sys/queue.h:451: note: this is the location of the previous definition
  451 | #define LIST_HEAD(name, type)                                           \
      |

-----

This is the fault of KallistiOS for including this header in its
public C standard library headers, but the only thing I can do about
it is get rid of this header in libxmp.

Realistically speaking, I think getting rid of this is fine for a few
reasons:

* Removing it makes iff.c slightly more readable.
* It is only used in iff.c. Past uses in load.c, ProWizard have since
  been removed since they relied on static globals.
* It is unlikely to be used in these locations again, since the
  loader and depacker lists are now static constant arrays.
* It is unlikely to be used for format allow/deny API functions
  (similar to MikMod) when/if these get added. The number of formats
  is known at compile time so these could just be implemented as
  fixed-size mirrors of the static constant lists. Linked lists offer
  essentially no benefits for this usage.

I've moved list.h to the regression tests, since it is probably safe
to keep using there.

This patch is fine to reject if it's not worth it.