Correct IFF chunk size type to uint32 for IFF loaders.
libxmp was inconsistently using long (internal to iff.c) and
int (loader functions) to describe the size of IFF chunks, when
this should really always be uint32 no matter what. I've corrected
this, though a couple places required hacks, and added a guard for
the absolute seek following chunk loading.

Issues with chunk sizes that this does NOT fix:
* Most loaders don't actually bother checking sizes.
* hio_seek doesn't support 64-bit values, so it can't correctly
  handle chunk sizes >INT_MAX except on 64-bit Linux et al.