libxmp_iff_load no longer fails on IFF container issues.
IFF issues (and really, individual chunk issues in general, except
in a few critical cases) should not result in entire modules being
rejected. I've changed most instances of IFF structural issues
causing loads to fail to instead act the same as the end of the
IFF file.

These cases are not actually that common (aside from the seek
failure on on 64-bit Windows and on 32-bit ); the most common
failure mode for broken IFF is to fail reading the chunk
signature, which was already treated as a successful load.