Add CoconizerSong loader.
Several Coconizer modules exist in a variant Coconizer "module"
format, in which the first readable string is "CoconizerSong".
This is not actually a magic string and this file format is
not actually a tracker module format, but rather, it is a RISC-OS
26-bit Relocatable Module (i.e. executable file) with an embedded
Coconizer module file somewhere inside of it. Several authors
distributed their modules only in this "format", which also
happens to contain two fields with unique copyright info and
comments not mirrored in their internal Coconizer modules.

Since this is a binary executable file, the most reliable way to
get the module offset is to search the start of the file for the
specific ARM instructions that the playercode uses to get the
module offset (which are fortunately very consistent between
versions). Once the module offset is verified, this loader can
just call into the normal Coconizer loader and extract the comments
when it's done.

This is probably the worst format that's been added to libxmp.
I am also writing a ripper for Modland, but this is still useful
for being able to detect and load these files in the wild without
extra manual effort.