read_event_ft2: finish cleaning up check instrument block.
```
SET(NEW_INS);
```
This is a flag that indicates an instrument number was in the event
usually. This flag seems to be mainly diagnostic and is only tested
in med_extras.c (irrelevant to here) and the invert loop handler
(also irrelevant here). The other handlers are varied about how
they actually set and clear this. The closest equivalent is the
default volume block, so moving it there.

```
xc->per_flags = 0;
```
XM, and the other formats that currently use `read_event_ft2`, do not
have any persistent effects implemented as far as I can tell, so this
is just here in case someone uses `-e`. The closest equivalent is the
default volume block, so moving it there.

```
if (!IS_VALID_INSTRUMENT(ev.ins - 1)) {
  /* If no note is set FT2 doesn't cut on invalid
   * instruments (it keeps playing the previous one).
   * If a note is set it cuts the current sample.
   */
```
Now explained elsewhere, safe to remove.

```
  xc->flags = 0;
```
Does nothing in this case except clears the newly set `NEW_INS`.
Since invalid instruments have pretty much the same semantics as
valid instruments for XMs, this flag isn't used for much, and
there's now precedence for *not* clearing this with invalid
instruments (`read_event_mod`), this can go.

```
  if (is_toneporta) {
    key = 0;
  }
}
```
The next usage of key is qualified with `!is_toneporta` and the usage
after that does `SET(NEW_NOTE)` (another mainly diagnostic flag that
is only used in med_extras.c), then does one of the following:

* `key == XMP_KEY_OFF`: the XM loader loads key off + ins# as
  `XMP_KEY_FADE`, so this case is irrelevant.
* `key == XMP_KEY_FADE`: is prevented from entering the old instrument
  block by the xyce-dans_la_rue.xm bugfix, so this case is irrelevant.
* `is_toneporta`: sets key to 0, which is the same thing that the
  instrument block does, so the instrument block piece can be removed.