FT2: don't reset fadeout for notes without an instrument.
Removes two problematic spots in read_event_ft2:

* Any case where a new note played, regardless of whether or not
  an instrument is present, would set the fadeout value to maximum.
  Prior to this line's most recent change in 09ad28be, this line was
  justified with a comment "for Last Battle.xm". This seems to be
  yet another workaround caused by handling FT2 keyoff cut with
  fadeout instead of channel volume.

* In the same block, if the current instrument is valid (this block
  is only reached in this case now) and doesn't have an envelope,
  the fadeout and release flags were reset too. This code has been
  revised several times (eb8b3a6f), also citing "Last Battle.xm".

The referenced Last Battle.xm is by Pixel and is from Cave Story.
It is not on Modland, probably because it's converted from Organya
(original at Organya/Pixel/cave story - 31 - last battle.org).
The affected sequence is in position 11, channel 2. The conversion
of this sequence is accurate to how the Organya original plays in
OrgMaker (and probably Cave Story, but I don't feel like replaying
it to the end just for this patch): the second pair of notes should
be quieter due to fadeout.

I fixed the test comment of test_player_ft2_note_noins_after_keyoff.c
too, since it was wrong about why Last Battle.xm cancels keyoff cut.